+++ CCL-Transfesa-CI_Order-Backend-Apirest-APP/src/main/java/com/ccl/transfesa/ci_order/backend/apirest/models/dao/IOrCiOrderDao.java
@@ -1,9 +1,8 @@
 package com.ccl.transfesa.ci_order.backend.apirest.models.dao;
 
-import java.math.BigDecimal;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import com.ccl.transfesa.ci_order.backend.apirest.models.entity.OrCiOrder;
 
-public interface IOrCiOrderDao extends PagingAndSortingRepository<OrCiOrder, BigDecimal> {
+public interface IOrCiOrderDao extends PagingAndSortingRepository<OrCiOrder, Long> {
 
 }
