+++ CCL-Transfesa-CI_Order-Backend-Apirest-APP/src/main/java/com/ccl/transfesa/ci_order/backend/apirest/models/service/IOrCiOrderService.java
@@ -1,6 +1,5 @@
 package com.ccl.transfesa.ci_order.backend.apirest.models.service;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.domain.Page;
@@ -14,10 +13,10 @@ public interface IOrCiOrderService {
 	
 	public Page<OrCiOrder> findAll(Pageable pageable);
 	
-	public OrCiOrder findById(BigDecimal id);
+	public OrCiOrder findById(Long id);
 	
 	public OrCiOrder save(OrCiOrder orCiOrder);
 	
-	public void delete(BigDecimal id);
+	public void delete(Long id);
 		
 }
